pyproject.toml
[project]
name = "novelvids"
version = "0.1.0"
description = "AI Novel Video Generator - Generate videos from novels using ComfyUI"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.12"
authors = [{ name = "NovelVids Team" }]
keywords = ["ai", "novel", "video", "comfyui", "tts", "generation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Video",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "tortoise-orm[asyncpg]>=0.22.0",
    "aerich>=0.8.0",
    "pydantic[email]>=2.10.0",
    "pydantic-settings>=2.6.0",
    "python-multipart>=0.0.18",
    "aiohttp>=3.11.0",
    "aiofiles>=24.1.0",
    "httpx>=0.28.0",
    "redis>=5.2.0",
    "celery>=5.4.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[argon2]>=1.7.4",
    "pillow>=11.0.0",
    "loguru>=0.7.0",
    "tenacity>=9.0.0",
    "uuid>=1.30",
    "transformers>=4.57.6",
    "torch>=2.9.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "pre-commit>=4.0.0",
]
sqlite = ["aiosqlite>=0.20.0"]
postgres = ["asyncpg>=0.30.0"]

[project.scripts]
novelvids = "novelvids.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["novelvids"]

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["novelvids"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "SIM", # flake8-simplify
]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["novelvids"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.aerich]
tortoise_orm = "novelvids.infrastructure.database.config.TORTOISE_ORM"
location = "./migrations"
src_folder = "."
